Moving vtgate's resolve methods to srvtopo.
